+++ R/wdpa_clean.R
@@ -19,7 +19,10 @@ NULL
 #'
 #' @param geometry_precision \code{numeric} level of precision for processing
 #'   the spatial data (used with \code{\link[sf]{st_set_precision}}). The
-#'   default argument corresponds to the nearest millimeter (i.e. 1000).
+#'   default argument is 1000 (higher values indicate higher precision).
+#'   This level of precision is generally suitable for analyses at the
+#'   national-scale. For analyses at finer-scale resolutions, please
+#'   consider using a greater value (e.g 10000).
 #'
 #' @param erase_overlaps \code{logical} should overlapping boundaries be removed
 #'   erased? This can be useful when the protected area boundaries are
@@ -161,7 +164,8 @@ wdpa_clean <- function(x,
                        crs = paste("+proj=cea +lon_0=0 +lat_ts=30 +x_0=0",
                        "+y_0=0 +datum=WGS84 +ellps=WGS84 +units=m +no_defs"),
                        snap_tolerance = 1,
-                       simplify_tolerance = 0, geometry_precision = 1000,
+                       simplify_tolerance = 0,
+                       geometry_precision = 1000,
                        erase_overlaps = TRUE,
                        verbose = interactive()) {
   # check arguments are valid
@@ -220,13 +224,12 @@ wdpa_clean <- function(x,
   x <- sf::st_set_precision(x, geometry_precision)
   x <- lwgeom::st_make_valid(x)
   x <- x[!sf::st_is_empty(x), ]
-  x <- rbind(suppressWarnings(sf::st_collection_extract(x, "POLYGON")),
-             x[vapply(sf::st_geometry(x), inherits, logical(1),
-                      c("POINT", "MULTIPOINT")), drop = FALSE])
+  x <- extract_polygons_and_points(x)
   if (verbose) {
     utils::flush.console()
     message("repairing geometry: ", cli::symbol$tick)
   }
+
   ## wrap dateline issues
   if (verbose) message("wrapping dateline: ", cli::symbol$continue, "\r",
                        appendLF = FALSE)
@@ -234,9 +237,7 @@ wdpa_clean <- function(x,
   x <- suppressWarnings(sf::st_wrap_dateline(x,
     options = c("WRAPDATELINE=YES", "DATELINEOFFSET=180")))
   x <- x[!sf::st_is_empty(x), ]
-  x <- rbind(suppressWarnings(sf::st_collection_extract(x, "POLYGON")),
-             x[vapply(sf::st_geometry(x), inherits, logical(1),
-                      c("POINT", "MULTIPOINT")), drop = FALSE])
+  x <- extract_polygons_and_points(x)
   x <- sf::st_set_precision(x, geometry_precision)
   if (verbose) {
     utils::flush.console()
@@ -248,9 +249,7 @@ wdpa_clean <- function(x,
   x <- sf::st_set_precision(x, geometry_precision)
   x <- lwgeom::st_make_valid(x)
   x <- x[!sf::st_is_empty(x), ]
-  x <- rbind(suppressWarnings(sf::st_collection_extract(x, "POLYGON")),
-             x[vapply(sf::st_geometry(x), inherits, logical(1),
-                      c("POINT", "MULTIPOINT")), drop = FALSE])
+  x <- extract_polygons_and_points(x)
   x <- sf::st_set_precision(x, geometry_precision)
   if (verbose) {
     utils::flush.console()
@@ -271,9 +270,7 @@ wdpa_clean <- function(x,
                        appendLF = FALSE)
   x <- lwgeom::st_make_valid(x)
   x <- x[!sf::st_is_empty(x), ]
-  x <- rbind(suppressWarnings(sf::st_collection_extract(x, "POLYGON")),
-             x[vapply(sf::st_geometry(x), inherits, logical(1),
-                      c("POINT", "MULTIPOINT")), drop = FALSE])
+  x <- extract_polygons_and_points(x)
   x <- sf::st_set_precision(x, geometry_precision)
   if (verbose) {
     utils::flush.console()
@@ -373,7 +370,7 @@ wdpa_clean <- function(x,
     message("formatting attribute data: ", cli::symbol$tick)
   }
   ## remove overlaps data
-  if (erase_overlaps) {
+  if (erase_overlaps && isTRUE(nrow(x) > 1)) {
     if (verbose) message("erasing overlaps: ", cli::symbol$continue)
     x$IUCN_CAT <- factor(as.character(x$IUCN_CAT),
                          levels = c("Ia", "Ib", "II", "III", "IV", "V", "VI",